+++ src/Advertiser/Service/StatsDataProvider.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright (c) 2018-2021 Adshares sp. z o.o.
+ * Copyright (c) 2018-2023 Adshares sp. z o.o.
  *
  * This file is part of AdServer
  *
@@ -32,12 +32,8 @@ use Adshares\Advertiser\Dto\Result\Stats\DataCollection;
 
 class StatsDataProvider
 {
-    /** @var StatsRepository */
-    private $repository;
-
-    public function __construct(StatsRepository $repository)
+    public function __construct(private readonly StatsRepository $repository)
     {
-        $this->repository = $repository;
     }
 
     public function fetch(StatsInput $input): StatsResult
@@ -46,14 +42,16 @@ class StatsDataProvider
             $input->getAdvertiserId(),
             $input->getDateStart(),
             $input->getDateEnd(),
-            $input->getCampaignId()
+            $input->getCampaignId(),
+            $input->getFilters(),
         );
 
         $data = $this->repository->fetchStats(
             $input->getAdvertiserId(),
             $input->getDateStart(),
             $input->getDateEnd(),
-            $input->getCampaignId()
+            $input->getCampaignId(),
+            $input->getFilters(),
         );
 
         return new StatsResult($total, $data);